Fix: App responsiveness.
- Made app more responsive.
- Fixed zooming inconsistencies.
- Fixed grid layout issues.
